token authentication middleware created

+++ 2-blog/controllers/blogs.js
@@ -1,15 +1,6 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
 const User = require("../models/user")
-const jwt = require("jsonwebtoken")
-
-const getWebToken = (request) => {
-  const authorization = request.get("authorization")
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "")
-  }
-  return null
-}
 
 blogsRouter.get("/", async (req, res) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
@@ -19,19 +10,16 @@ blogsRouter.get("/", async (req, res) => {
 blogsRouter.post("/", async (req, res) => {
   const body = req.body
 
-  const decodedToken = jwt.verify(getWebToken(req), process.env.SECRET)
-  if (!decodedToken.id) {
-    return res.status(401).json({ error: "token missing or invalid" })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(body.userId)
 
   const newBlog = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: user.id,
+    user: user._id,
   }
+
   if (newBlog.likes === undefined) {
     newBlog.likes = 0
   }
@@ -54,6 +42,13 @@ blogsRouter.post("/", async (req, res) => {
 
 blogsRouter.delete("/:id", async (req, res) => {
   const id = req.params.id
+  const user = await User.findById(req.body.userId)
+  const blog = await Blog.findById(id)
+  if (!blog || !user || user._id.toString() !== blog.user.toString()) {
+    return res
+      .status(400)
+      .send({ error: "only the creator of this blog can delete it" })
+  }
   await Blog.findByIdAndRemove(id)
   res.status(204).end()
 })
